refactored BaseSequencer to have several files by kind of movement
